+++ packages/shared/akamai-edgeworker-sdk/CHANGELOG.md
@@ -80,6 +80,12 @@ All notable changes to the LaunchDarkly SDK for Akamai Workers will be documente
   * dependencies
     * @launchdarkly/js-server-sdk-common bumped from ^2.1.1 to ^2.1.2
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common bumped from ^2.2.1 to ^2.2.2
+
 ## [1.1.1](https://github.com/launchdarkly/js-core/compare/akamai-edgeworker-sdk-common-v1.1.0...akamai-edgeworker-sdk-common-v1.1.1) (2024-03-15)